Trigger refresh of Click to Load placeholders on request block
Historically, the Click to Load placeholders were only drawn once at
the point page load completed. That had two problems:
1. Sometimes placeholders would take too long to display, if the page
   took a long time to load fully.
2. Sometimes placeholders would not be displayed at all, if the
   blocked content was only created after page load had finished.

To fix that, let's send a message to the content-scope-script to
ensure that placeholders are refreshed each time a Click to Load
request is blocked.

Also tweak the response messages to be inline with the recent
content-scope-script changes to make response messages easier to
identify.

Unfortunately the YouTube Click to Load integration tests need to be
disabled for now, since we had to disable the "hideTrackingElement"
option for now which breaks some of the YouTube SDK integration.
