+++ integration-test/config.json
@@ -2,6 +2,7 @@
     "spec_dir": "integration-test",
     "spec_files": [
         "legacy/*.js",
-        "content-scripts/gpc.js"
+        "content-scripts/gpc.js",
+        "!legacy/click-to-load-youtube.js"
     ]
 }
