+++ shared/js/content-scripts/content-scope-messaging.js
@@ -19,21 +19,34 @@ function getSecret () {
 async function init () {
     const secret = await getSecret()
 
+    // Content-scope-script messaging proxy, to allow the Click to Load content
+    // script to send messages to the extension's background and receive a
+    // response.
+    // Note: This event listener is only for Chrome MV3 builds of the extension,
+    //       the equivalent Chrome MV2 event listener lives in
+    //       content-scope-scripts/inject/chrome.js.
     window.addEventListener('sendMessageProxy' + secret, event => {
-        // MV3 message proxy for click to load
         event.stopImmediatePropagation()
-        const detail = event && event.detail
-        if (!detail) {
+
+        if (!(event instanceof CustomEvent) || !event?.detail) {
             return console.warn('no details in sendMessage proxy', event)
         }
-        const messageType = detail.messageType
+
+        const messageType = event.detail?.messageType
         if (!allowedMessages.includes(messageType)) {
             return console.warn('Ignoring invalid sendMessage messageType', messageType)
         }
-        chrome.runtime.sendMessage(detail, response => {
-            const msg = { type: 'update', detail: { func: detail.messageType, response } }
+
+        chrome.runtime.sendMessage(event.detail, response => {
+            const message = {
+                type: 'update',
+                messageType: 'response',
+                responseMessageType: messageType,
+                response
+            }
+
             window.dispatchEvent(new CustomEvent(secret, {
-                detail: msg
+                detail: message
             }))
         })
     })
